requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
future==0.18.3
idna==3.4
load-dotenv==0.1.0
netmiko==4.1.2
ntc-templates==3.2.0
numpy==1.24.2
pandas==1.5.3
paramiko==3.0.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scp==0.14.5
six==1.16.0
tenacity==8.2.0
textfsm==1.1.2
urllib3==1.26.14
